requirement.txt
requests
python-dotenv
langchain
langchain-openai
pandas
scikit-learn
matplotlib
google-search-results
langchain_community
langchain.tools
bs4